Work-in-progress: settings split and lazyload
