+++ server/scripts/proxy.js
@@ -33,7 +33,8 @@ app.use(
             "/ushahidi-legacy-app.js",
             "/0.ushahidi-legacy-app.js",
             "/1.ushahidi-legacy-app.js",
-            "/activity.js"
+            "/activity.js",
+            "/settings.js"
         ],
         {
             target: `http://localhost:${LEGACY_PORT}/`,
